+++ pages/index.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from "next";
 import Head from "next/head";
 import Image from "next/image";
 import styles from "../styles/Home.module.css";
-import ProductPage from "./productPage";
+import ProductPage from "./item";
 
 const Home: NextPage = () => {
   return <ProductPage />;
